version 0.5.5

+++ simsce.h
@@ -2,7 +2,7 @@
 #define SIMSCE_H
 #include <QString>
 
-static int filePieceNum ;//文件块数
+static double filePieceNum ;//文件块数
 static double seedUpLink;//种子节点上传带宽
 
 class SimSce
